Give inline methods hidden visibility by default
Hidden inline methods will not appear in the export table of a DSO
(dynamic shared object) and will not require a PLT indirection when used
within the DSO.

According to the GCC man page:
Enabling  this  option  can have a dramatic effect on load and link
times of a DSO as it massively reduces the size of the dynamic export
table when the library makes heavy use of templates.

So:
- it is optimization
- also hides our internal inline methods

And we can do this because we do not attempt to compare pointers
to inline functions or methods where the  addresses  of the two
functions are taken in different shared objects.