system-upgrade: backwards-compat offline-{upgrade,distrosync}
Make `dnf5 offline-distrosync` and `dnf5 offline-upgrade` backwards-compatible with
DNF 4. There didn't seem to be a way to do this with aliases alone since
we need `dnf5 offline-distrosync` and `dnf5 offline-distrosync download`
to be aliased to different commands (`dnf5 offline` and `dnf5
distro-sync`, respectively). We also want `dnf5 offline-distrosync --help`
to show the help for `dnf5 offline`.