Fix `DISTRO_SYNC_ALL` (distro-sync without arguments, system upgrade)
Since distro-sync uses `SOLVER_TARGETED` mode we cannot pass in
installed packages because if we did updating only a subset of packages
would be a valid solution. However in a distro-sync we want to ensure
ALL packages are synchronized with the target repository.

This brings dnf5 behavior closer to dnf4.
There is still one difference in the transaction setup, dnf4 doesn't
specify the transaction as targeted but thanks to libsolv's
`auto-targeting` this should have the same behavior.

For: https://bugzilla.redhat.com/show_bug.cgi?id=2273749