dnfdaemon: Catch timeout during CheckAuthorization
- set a timeout for the user to enter the password (default 25 seconds
  seems to be too short)
- catch timeout error - otherwise the error is returned to the client
  causing the original call to fail with timeout

For example:

The user invokes the do_transaction() method on the Goal interface. The
first action of the do_transaction handler is to verify that the user is
authorized to run an RPM transaction. If an authentication agent is
available, this may result in prompting the user to enter their
password. If the user ignores the prompt, after a timeout period, the
result should be a "Not authorized" error.

However, without this patch, the result would be a "Method call timed
out," which is misleading. While the user called do_transaction(), the
timeout did not occur at this stage.