Base: notify_repos_configured and are_repos_configured methods
`notify_repos_configured` notifies the libdnf5 library that the repositories
are configured.  It can be called before `load_repos`. The libdnf5 library
can then call plugins that can make final adjustments to the repositories
configuration. In the case that it has not been called, it is called
automatically at the beginning of the `load_repos` method.