system-upgrade: adapt to new transaction serialization format
System-upgrade was broken by 5f8b76ad7e0faa5c5b151e4301417f004bc4470e,
since RPMs of serialized transactions are now expected to be stored in a
single flat directory (like that created by `dnf5 download --destdir
./packages`) rather than the directory structure used by the DNF cache
(in `/path/to/repository/cachedir/packages`)

We would like to set `destdir` to download the packages to a single flat
directory, but then we can't test the transaction as usual because of
the limitation described in
https://github.com/rpm-software-management/dnf5/pull/1301. So instead,
we serialize the transaction without testing it, then build a new "test
transaction" from the serialized transaction and test that. This
solution may seem like a hack, but it's actually quite nice, since we
can a test the offline transaction exactly how it will be performed
later by `dnf5 offline _execute`.

Due to the improvements from
https://github.com/rpm-software-management/dnf5/pull/1264, we can now
create the offline transaction without making a copy of the entire
repository cache, and we can execute it without loading the repositories
at all.