conf: New classes for append options
Introducing `OptionStringAppendList` and `OptionStringAppendSet` classes for
append options.

The current implementation of append options is flawed because it
requires the options to be set in ascending order of priority for proper
functionality. This is necessary for the resetting mechanism using an
empty item to work correctly. However, in dnf5, options are not
processed strictly in priority order. For instance, command line options
like `--exclude` are processed before the configuration files, despite
having a higher priority than config files.

The new classes accumulate all setting attempts and then process them in
priority order to compute the actual value.