Added de-duplication for environment list print
Previously, `dnf5 environment list` would print duplicate environment
with the same id and name. It would be confusing for the user to see
duplicate environment names. This patch introduces de-duplication for
the environment list print feature to merge duplicate environment
entries. The original order of the environments is preserved by
assigning an index value to each environment entry.

Fixes: #1502