Add XML output option to 'dnf5 advisories list'
Add an option to output the list of applicable advisories in the same
XML format as the updateinfo.xml file in the repository.

This allows tooling to have a consistent machine readable format for
getting what advisories apply to a system. Resorting to parsing text
output of commands is (naturally) error prone, and designed for humans
to look at rather than machines.

Alternatives for these tools include using the APIs of
dnf/yum/other-rpm-package-manager, and keeping up with their evolution.
For tooling that supports a wide range of distributions and versions,
this can become rather onerous, especially across the many
different security/patch management products.

With a documented format, we can help these tools get the correct
information both on available updates, and what ones are applicable to
particular systems. Using the same format these tools need to be able to
parse already avoids the need to get multiple parsers correct.

Signed-off-by: Stewart Smith <trawets@amazon.com>