Add JSON output to advisory list
Similar to the repolist --json output, do it for listing advisories that
apply to the system ("dnf advisory list") as well as for the references
currently supported Bugzilla and CVE types of references.

These follow what is output on the terminal. The exception is that the
queries with a reference (--with-bz and --with-cve) also print the
advisory name. This is just to make it a bit easier for other tools
which may already know about that advisory ID.

Example of advisory list --json:
[
  {
    "name":"FEDORA-2024-143f0443e1",
    "type":"unspecified",
    "severity":"None",
    "nevra":"yum-4.19.2-1.fc40.noarch",
    "buildtime":"2024-04-19 21:20:20"
  },
  {
    "name":"FEDORA-2024-1e8a70c30e",
    "type":"enhancement",
    "severity":"None",
    "nevra":"zstd-1.5.6-1.fc40.aarch64",
    "buildtime":"2024-04-19 21:20:20"
  }
]

Example of advisory list --with-bz --json:
[
  {
    "advisory_name":"FEDORA-2024-0c9d3b51d4",
    "advisory_type":"security",
    "advisory_severity":"security",
    "advisory_buildtime":"2024-05-02 01:55:56",
    "nevra":"tpm2-tss-fapi-4.1.0-1.fc40.aarch64",
    "references":[
      {
        "reference_id":"2271763",
        "reference_type":"bugzilla",
      },
      {
        "reference_id":"2277437",
        "reference_type":"bugzilla",
      }
    ]
  }
]