Implement conditional compilation `-DWITH_MODULEMD=OFF`
When WITH_MODULEMD is turned off module sources are not compiled and
module headers are not installed.

Public API in non-module headers (base.hpp and goal.hpp) is still
available but throws an exception when used.
Another option would be to also hide the functions in non-module headers
but in order to ensure the headers always match the compiled library (so
they don't depend on the user not defining `WITH_MODULEMD`) they would
have to be configured via a constant at compile time. For example using
cmake to `configure_file` turning them into `.in` files.