Fix DNF5: Don't trigger filelists download if abs path to local RPM
If the path to the local package was specified absolutely (starting with
'/'), filelists download was triggered.
Now, as in `Goal`, DNF5 checks if the path ends with the ".rpm" suffix.
If it does, it is treated as the path to the local RPM package and
filelists download is not triggered.