Use `SOLVER_FLAG_FOCUS_NEW` to install latests versions of deps
This is a new flag in the recent 0.7.30 release.
It is used only in current rawhide.

We previously tried to use `SOLVER_FLAG_FOCUS_BEST` but it turned out to
be too strong, this new flag should update packages only if it is
required for installing the latests versions of dependencies.

We had several reports where users expect this behavior.

This should properly resolve issue such as:
https://bugzilla.redhat.com/show_bug.cgi?id=2150300
https://issues.redhat.com/browse/RHEL-1374