When upgrading a group keep stored package types
This matches the dnf4 behavior.

It fixes confusing behavior where when a `group upgrade` is executed on a new
version of a group with new packages it installs only the new packages that
match the stored (in system state) package types but it also
overrided the package types with the currently configured `group_package_types`.