Add `override_reasons` to `GoalJobSettings` and use it in `redo`
If a package is already installed in the requested version but with a different reason override the reason.
This is useful when redoing an interrupted transaction.
The longest part of a transaction is running rpm changes, saving reasons is done after rpm finishes. If a transaction
is interrupted in the middle of the rpm part some rpms are changed but reasons are not updated at all. To handle this
enforce the reasons in the selected transaction.