spec: Stricten a dependency on DNF libraries in plugin subpackages
ELF plugins of dnf5-plugins and dnf5-plugin-automatic subpackages link
to libdnf and libdnf-cli libraries. Because the libraries come from
the same source package, it is desired to pin the plugin subpackages
to same release of the libraries subpackages.

This was hinted by rpminspect and already partially implemented in
commit 97cdbe57c34259a0e0dc87431ac7c3e9a396aba7 ("Add explicit package
version for libdnf5-cli").