cmdline_aliases: update to `toml11-devel-4.0.0` API
Adds missing cmake check for toml11.

It preserves the old API use with conditional compilation based on the
found toml11-devel version.

Uses `toml::ordered_value` to keep the order of values from files.
https://toruniina.github.io/toml11/docs/features/configure_types/#preserving-order-of-values-in-tables
(It now uses `toml::ordered_map` instead of `libdnf5::PreserveOrderMap`
they do the same thing but the libdnf5 version is missing a constructor).
It now preserves comments but they are not used and it shoudn't have an
effect.

The `source_location` API was changed.
https://toruniina.github.io/toml11/docs/reference/source_location/

It simplifies formatting of `syntax_error`s.