state.cpp: update to new toml11-devel-4.0.0 version API
Adds TC - `type_config` which is required for conversion.
https://toruniina.github.io/toml11/docs/features/value/#converting-to-tomlvalue

The `toml::format` was changed and no longer accepts template args:
https://toruniina.github.io/toml11/docs/reference/serializer/

Use `ordered_value` to preserve order, otherwise `std::unordered_map` is
used internally which may result in different order on each run which
may casue failing tests.

It was necessary to specify the formating for some values because the
tests rely on it. However even with the same formatting there still are
changes (spacing is different), update the tests to pass for both toml11
versions.