dnfdaemon: Signals should use "object_path" type for the session_object_path argument
The interface is defined as the "session_object_path" is of type "object_path",
but when setting the arguments it's passing the path as a string, which causes
a type mismatch and the client side can reject the signal.

Store the object path as an sdbus::ObjectPath in the Session class to
have correct types in the signal emission.

Closes https://github.com/rpm-software-management/dnf5/issues/1646