RepoSack: New API to reload system repo
Introduce a new API to clear all solvables from the system repo and
reload fresh data from the RPM database.

Re-creating the Base object can be resource-intensive, and there is a
use case in dnfdaemon where the current session (which wraps around the
Base instance) needs to be reused even after executing an RPM
transaction - whether for queries or running another transaction.