When writing main solv file (primary.xml) don't store filelists
In a repo filelists are stored in different repodata, limit the written
repodata when writing main solv file.

Otherwise the filelists are stored twice: in the main solv file and in the
filelists specific solvx file. (This happens when downloading both the
primary.xml and filelists.xml at the same time.)

It also improves performance because when filelists were part of main
solv file they were always loaded (regardless of the
`optional_metadata_types` option).