Recommend --use-host-config if --installroot is used and not all repositories can be enabled
It was reported that --installroot with an explicit --repo option
does not recognize given repository despite the repository is defined
under /etc:

    $ dnf --installroot /tmp/dnf5/ --releasever rawhide --repo koji makecache
    No matching repositories for "*, koji". Add "--help" for more information about the arguments.

That's a feature, but a difference from DNF4.

This patchs recommeds using --use-host-config in that case. We already
have a similar recommendation in main.cpp in case of no --repo option is used.

This patch also fixes formatting of the list of repository
identifiers with respect to localization.

The code duplicates throws because libdnf5 exception construtors do
not accept nonstatic arguments.

Resolves #1756