Do not depend on the repository's pot file when generating translations
Currently, dnf5 stores the pot (translation template) file in the dnf5
repository, as well as in the dnf5-l10n repository. The dnf5-l10n
repository's pot files are regularly updated by a command from the dnf5
repository. However, the dnf5 repository's pot file is not updated
automatically.

While recent commits have updated the pot file in the dnf5 repository
during the release process, it seems keeping the pot file in
the repository isn't necessary. We can generate the pot file using
gettext's xgettext command, and we already have a CMake target
to generate the pot file.

This commit changes the destination of the pot file generated by the
xgettext command to the build directory. Removing the pot file from
the repository will be done in the next commit.