Optimize getting counts of transaction items
Previously during `history list` we separately fetched all transaction
packages, groups and environments only to count them. This could be
quite slow because it resulted in a lot of separate DB connections.

Now we fetch only the counts and use only one DB connection.
To maintain compatibility the counts are passed as an optional parameter
and if missing we fallback to the old (slow) computation.