DownloadCallbacks: Ensure `end` for every successful `add_new_download`
For every successful (ended without exception) call of
`add_new_download` callback, the `end` callback is called once.

The code relied on librepo to call the `end` callback just once for each
`lr_target`. Which works under normal conditions. But if there is
an interruption in the librepo download (error, signal received) librepo
will not call `end` callbacks for targets that have not yet started downloading.
However, the `add_new_download` callback has already been called for them.
For example, if a higher layer allocated resources on `add_new_download` and
deallocated on `end`, resources would be lost.

The new code tracks this and calls the appropriate `end` callbacks.
Also, an assert has been added to monitor for any attempt to double call
the `end` callback.