libdnf5: Add a plugin to download and install repo's Appstream data
Repositories can provide Appstream data for the packages they contain.
This Appstream data is consumed by applications like the GNOME Software
or the KDE Discover, thus the users can see the packages (apps) in them.

This is to be in pair with PackageKit, which does download and install
the repo's Appstream data.

The plugin is built by default, but the Appstream data is not downloaded
unless "optional_metadata_types" config option contains "appstream".

The plugin adds a dependency on the `appstream` library. It can be disabled
with a WITH_PLUGIN_APPSTREAM CMake option set to OFF.

Closes https://github.com/rpm-software-management/dnf5/issues/1564