automatic: add a default setting to not emit boring messages
Currently, we always emit messages every time automatic runs,
even if it does nothing very interesting (the operation succeeds
and there are no changed packages). This is flooding Fedora infra
with boring emails that effectively just say "nothing happened".

This adds a setting that controls whether we emit a message when
nothing interesting happened, and makes it default to false, so
we only emit if *something* interesting happened - the operation
failed, or we saw package changes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>