Fix reporting disk space to be freed on a pure package removal
There was a bug in testing whether a difference of an installation size
and a removal size fits into int64_t. E.g. on a pure package removal,
the difference was negative, overflowing resulting unsigned type,
leading to a very large positive value, larger than could fit into
int64_t. As a result an existing range check failed and no report was
printed.

This patch fixes it by preventing from a subtraction with a negative
rusult.

(A fix which would compute the difference before the branching
would either loose precision, or decide the branching twice.)

Resolves: #1938