Add --json output to advisory info
Enable --json for 'dnf advisory info'. The same information as would be
displayed as text is included in the JSON output. While the libscols
code could be adapted to produce JSON as well, without serious reworking
of key_value_table (and all callers), the produced JSON is quite weird,
so we instead choose to implement the JSON output separately.