Bash completion: always offer NEVRAs for packages
Originally, for packages, the NAME was offered in bash completion.
In the case of multiple packages with the same NAME, there was an option
to offer NEVRA instead of the NAME for those packages.

Now, bash completion always offers NEVRA for packages.
There is an option to offer package NAMEs in addition to NEVRAs. This is
useful, for example, when the user wants to install a package. Usually
the user only wants to enter the package NAME and not the entire NEVRA.
If multiple packages with the same NAME are available, the specific
package is then selected by the solver.