For packages, always offer nevra in bash completion
Originally, for packages, the name was offered in bash completion.
In the case of multiple packages with the same name, there was an option
to offer navra instead of a name for those packages.

Now, bash completion always offers nevra for packages. In the case
of multiple packages with different nevras but the same name, there is
an option to offer a name in addition to nevras for those packages.

For example, a user wants to install the package "xyz". The package
is available in multiple versions. Bash completion will offer
the nevras of all versions plus just the name. Usually the user
does not install a specific nevra, but the name and the solver selects
the version.