Make `history info` without trans ID more user friendly and dnf4 compatible
When the default value of the transaction ID argument is only `last`
users searching for transactions with specific packages might be confused
why the output is empty.

For example:
```
$ dnf5 history info --contains-pkgs=htop
```
Would search only the last transaction. On the other hand `list` works as
expected:
```
$ dnf5 history list --contains-pkgs=htop
```
Searches all transactions becuase the default for `list` is all
transactions.

This adds additional logic to use the `last` transaction only if
`--contains-pkgs` is empty.