Add a hint to `history info` without trans IDs when no match found
When the default value of the transaction ID argument is only `last`
users searching for transactions with specific packages might be confused
why the output is empty.

For example:
```
$ dnf5 history info --contains-pkgs=htop
```
Would search only the last transaction. On the other hand `list` works as
expected:
```
$ dnf5 history list --contains-pkgs=htop
```
It searches all transactions becuase the default for `list` is all
transactions.