Copr plugin: Fix resource leak in load_all_configuration
The "globfree" function call was missing.

One can say that adding the line `globfree(&glob_result);` after the for
loop is enough. But it's not true. The `load_copr_config_file` method
called in the loop can generate an exception. So I rewrote the code
to use `std::filesystem::directory_iterator`.