Optimize build process: No libdnf5_static.a object files archive
The "libdnf5_static.a" object files archive was not packaged. It was
only used in unit tests of the "libdnf5" library. The use of the archive
in the "libdnf5" unit tests was intentional. The unit tests call private
methods that have hidden symbols not exported by the shared library.

Now the `INTERFACE` library is defined instead of the object files
archive. It specify usage requirements for dependents but does not
produce a library artifact on disk. The "libdnf5" unit tests link
directly to the object files without the need for an intermediate
archive.