DNF5: Support for suggesting command line args without description
If the argument at position 1 is
`--complete=<index>[,add_description=1/0]`, this is a request
to complete the argument at position `<index>`.

Before, only `--complete=<index>` was used and the following text was
ignored. Now the optional `,add_description=1/0` is recognized, which
enables/disables the addition of a description to the proposed command
line arguments. By default, adding a description is enabled.