DNF5 bash completion: "menu completion" support
"Menu completion" cyclically completes the suggestions from the list.
We don't want to add a description to the added argument -> We need
a list of suggestions without descriptions.

"Normal completion" completes an identical part of the suggestions.
The added description is not printed. -> Descriptions can be
in the suggestion list, but will not be used. We can optimize the code
and not add the descriptions to the list.