load_plugins: Preserve original exception with failure information
The original implementation threw an exception with a list of plugins
that failed to load. We usually got the name of one configuration file.
Information about the reason for the failure was not included
in the exception.

Now the code throws an exception for the first failed plugin. Exception
contains the full path (not just the filename) and includes the original
exception from the source of the error.