RepoSack::update_and_load_repos: Properly thread_sack_loader termination
The `OnScopeExit` class is used to ensure proper termination of
the `thread_sack_loader` in case of unexpected exit of
the `update_and_load_repos` method due to an exception.

If the `thread_sack_loader` is not terminated (joined) before the method
exits, the program will crash. Before this fix, the program crashed,
for example, on an exception generated during pgp key import.