Actions plugin: Create pipes with close-on-exec flag
The pipelines are closed manually. But there is a risk that
an application that uses libdnf will create a child process (`fork`) in
another thread and execute a program in it. This will leak open pipe
descriptors into the newly executed program.
The close-on-exec flag (`O_CLOEXEC`) ensures that the descriptors are
not passed to the newly executing program (they are closed).