Create progress bar in script callback if one doesn't already exist
Normally, the script start callback is called after package verification
has been done, so there is an active progress bar (created in the
verification stage), however package verification can be disabled. If
verification is disabled, then the script start callback is the first
callback used (to run pre-trans scripts), and there will be no active
progress bar, leading to a crash.

To handle the case of disabled verification, create a new progress bar
in the script start callback for it to use in the future.