Actions plugin: Introduce `raise_error` option
A new optional boolean option `raise_error` was added. The allowed
values are "0" and "1". For backwards compatibility, the default value
is "0".

If `raise_error` is set to "1", an exception is thrown if the action
process failed to start or ended with a non-zero exit code or ended
abnormally (received a signal) an error occurred during communication
(syntax error, communication interrupt, failed to process the output
line in plain communication mode).