libdnf5 plugin API: Add `StopRequest` class
Plugin hooks do not return a value (void methods). Exceptions are used to pass
errors from hooks.

The `StopRequest` class is used to mark a plugin-generated exception as a stop
request. The stop request is not caused by an error in the plugin runtime, but
is a plugin request. The exception is marked by inheriting from
the `StopRequest` class (using multiple inheritance).

The libdnf5 plugin API change is backwards compatible. Only the minor version
of the API is increased: "2.1" -> "2.2"