Create a pipe and open files with the close-on-exec flag
An application that uses libdnf can create a child process (`fork`)
in another thread and execute the program in it.  This will leak open
descriptors to the newly executed program.

The close-on-exec flag (`O_CLOEXEC`) ensures that the descriptors are
closed and not passed to the newly running program (they are closed).