Fix RepoCache::Impl::remove_recursive: Do not follow symlinks
Symlinks were followed when removing the libdnf5 cache. This is very
dangerous.
Example:
A user extracted a package into the cache folder. And later runs
`dnf clean all`. The package contained a recursive symlink, which
terminated the deletion with
a `filesystem error: status: too many levels of symbolic...` error.

Imagine what would happen if there was a symlink outside the cache. Like
a link to "/home", "/"...