Fix total number of transaction progress bars
Before the transaction, the total number of steps needed and thus
the progress bars needed is calculated. The original calculation assumed
that if new packages enter the system through transactions, the first
step would be to verify them. However, package verification can be
disabled.

This modification will add the verification step to the total number
of steps only if verification is performed.