SWIG: Implement exception forwarding
Exceptions were passed as `SWIG_RuntimeError`. The original exceptions type
was dropped and the information contained too. Only the `what()` message was
retained. Nested exceptions were completely ignored.

Now the original exceptions, including nested exceptions, are passed to
the target language. The exceptions implement the `rethrow_if_nested` method,
which can be used to raise a nested exception if it is present. And the `format`
method, which formats the exception error message, including the error messages
of nested exceptions.